Move MC specific hdf5 code to hdf5::mc
