+++ src/io/mcreaders.rs
@@ -14,24 +14,6 @@ mod grr {
 
     #[derive(hdf5::H5Type, Clone, PartialEq, Debug)]
     #[repr(C)]
-    pub struct Vertex {
-        pub event_id: u32,
-        pub track_id: u32,
-        pub parent_id: u32,
-        pub x: f32,
-        pub y: f32,
-        pub z: f32,
-        pub t: f32,
-        pub moved: f32,
-        pub pre_KE: f32,
-        pub post_KE: f32,
-        pub deposited: u32,
-        pub process_id: u32, // NB these may differ across
-        pub volume_id: u32,  // different files
-    }
-
-    #[derive(hdf5::H5Type, Clone, PartialEq, Debug)]
-    #[repr(C)]
     pub struct SensorHit {
         pub event_id: u32,
         pub sensor_id: u32,
@@ -45,18 +27,6 @@ mod grr {
         pub sensor_id: u32,
         pub charge: u32,
     }
-}
-
-#[derive(hdf5::H5Type, Clone, PartialEq, Debug)]
-#[repr(C)]
-pub struct Primary {
-    pub event_id: u32,
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
-    pub vx: f32,
-    pub vy: f32,
-    pub vz: f32,
 }
 
 #[derive(hdf5::H5Type, Clone, PartialEq, Debug)]
@@ -87,16 +57,8 @@ pub struct SensorReadout {
     pub t: Time,
 }
 
-pub fn read_vertices(filename: &Path, range: Bounds<usize>) -> hdf5::Result<Vec<Vertex>> {
-    Ok(array_to_vec(read_table::<Vertex>(&filename, "MC/vertices", range)?))
-}
-
 pub fn read_sensor_hits(filename: &Path, range: Bounds<usize>) -> hdf5::Result<Vec<SensorHit>> {
     Ok(array_to_vec(read_table::<SensorHit>(&filename, "MC/waveform", range)?))
-}
-
-pub fn read_primaries(filename: &Path, range: Bounds<usize>) -> hdf5::Result<Vec<Primary>> {
-    Ok(array_to_vec(read_table::<Primary>(&filename, "MC/primaries", range)?))
 }
 
 pub fn read_sensor_xyz(filename: &Path) -> hdf5::Result<Vec<SensorXYZ>> {
@@ -148,28 +110,11 @@ mod test_mcreaders {
     }
 
     #[test]
-    fn test_read_vertices() {
-        let test_file = get_test_file();
-
-        let vertices = read_vertices(test_file, Bounds::none());
-        let track_ids: Vec<u32> = vertices.unwrap().into_iter().map(|vrt| vrt.track_id).collect();
-        assert_eq!(track_ids, [1, 1, 2, 4, 1]);
-    }
-
-    #[test]
     fn test_read_sensor_hits() {
         let test_file = get_test_file();
 
         let sensor_hits = read_sensor_hits(test_file, Bounds::none());
         assert_eq!(sensor_hits.unwrap().len(), 8098);
-    }
-
-    #[test]
-    fn test_read_primaries() {
-        let test_file = get_test_file();
-
-        let primaries = read_primaries(test_file, Bounds::none());
-        assert_eq!(primaries.unwrap().len(), 5);
     }
 
     #[test]
